requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0